Move article with native provider to pages package
