+++ android-app/index.android.js
@@ -1,10 +1,13 @@
 import { AppRegistry } from "react-native";
-import AuthorProfileView from "./pages/author-profile";
-import ArticleView from "./pages/article";
-import SimpleArticleView from "./pages/simple-article";
-import TopicView from "./pages/topic";
+import AuthorProfileView from "./src/pages/author-profile";
+import ArticleView from "./src/pages/article";
+import ArticleWithNativeProvider from "./src/pages/article-with-native-provider";
+import TopicView from "./src/pages/topic";
 
 AppRegistry.registerComponent("AuthorProfile", () => AuthorProfileView);
 AppRegistry.registerComponent("Article", () => ArticleView);
-AppRegistry.registerComponent("SimpleArticle", () => SimpleArticleView);
+AppRegistry.registerComponent(
+  "ArticleWithNativeProvider",
+  () => ArticleWithNativeProvider
+);
 AppRegistry.registerComponent("Topic", () => TopicView);
