+++ android-app/package.json
@@ -15,7 +15,6 @@
   },
   "dependencies": {
     "@times-components/pages": "1.5.3",
-    "@times-components/provider": "1.3.1",
     "prop-types": "15.6.2",
     "react": "16.4.2",
     "react-native": "0.55.4",
