+++ android-app/src/ad-config.js
@@ -1,0 +1,16 @@
+export default (config, sectionName) => ({
+  adUnit: "d.thetimes.co.uk",
+  networkId: "25436805",
+  testMode: "",
+  appVersion: config.appVersion,
+  operatingSystem: "Android",
+  operatingSystemVersion: config.operatingSystemVersion,
+  environment: config.environment,
+  deviceId: config.deviceId,
+  cookieEid: config.cookieEid,
+  cookieAcsTnl: config.cookieAcsTnl,
+  cookieIamTgt: config.cookieIamTgt,
+  isLoggedIn: config.isLoggedIn,
+  platform: "mobile",
+  sectionName
+});
