+++ android-app/src/article-prop-types.js
@@ -1,0 +1,8 @@
+import PropTypes from "prop-types";
+
+export default {
+  articleId: PropTypes.string.isRequired,
+  omitErrors: PropTypes.bool.isRequired,
+  scale: PropTypes.string.isRequired,
+  sectionName: PropTypes.string.isRequired
+};
