+++ packages/pages/src/article-prop-types.js
@@ -1,0 +1,22 @@
+import PropTypes from "prop-types";
+
+const propTypes = {
+  analyticsStream: PropTypes.func.isRequired,
+  platformAdConfig: PropTypes.shape({}).isRequired,
+  omitErrors: PropTypes.bool,
+  onArticlePress: PropTypes.func.isRequired,
+  onAuthorPress: PropTypes.func.isRequired,
+  onCommentsPress: PropTypes.func.isRequired,
+  onCommentGuidelinesPress: PropTypes.func.isRequired,
+  onVideoPress: PropTypes.func.isRequired,
+  onLinkPress: PropTypes.func.isRequired,
+  onTopicPress: PropTypes.func.isRequired,
+  scale: PropTypes.string.isRequired,
+  sectionName: PropTypes.string.isRequired
+};
+
+const defaultProps = {
+  omitErrors: false
+};
+
+export default { propTypes, defaultProps };
