+++ packages/pages/src/pages.js
@@ -1,4 +1,6 @@
 export { default as Article } from "./article";
-export { default as SimpleArticle } from "./simple-article";
+export {
+  default as ArticleWithNativeProvider
+} from "./article-with-native-provider";
 export { default as AuthorProfile } from "./author-profile";
 export { default as Topic } from "./topic";
